Abstract out standard config
This also fixes that we weren't extending recommendedTypeChecked, which
meant a lot of unnecessary type assertions getting cleaned up.
