+++ play.pokemonshowdown.com/src/battle-text-parser.ts
@@ -1214,5 +1214,5 @@ declare const require: any;
 declare const global: any;
 if (typeof require === 'function') {
 	// in Node
-	(global as any).BattleTextParser = BattleTextParser;
+	global.BattleTextParser = BattleTextParser;
 }
